Split runner apply text edits tests
